moved _git method and dependencies to tool.py, since it's needed only for testing now
